requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
click==8.3.0
colorama==0.4.6
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
scramp==1.4.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
